Create Error & Solution: The code must be changed so the plugin only appears in one plugin file. You will need to remove the macro which registers the plugin so it only appears in one of these files.   If none of these files register such a plugin, then the problem originates in a library to which all these files link. The plugin registration must be removed from that library since plugins are not allowed in regular libraries.
The code must be changed so the plugin only appears in one plugin file. You will need to remove the macro which registers the plugin so it only appears in one of these files.
  If none of these files register such a plugin, then the problem originates in a library to which all these files link.
The plugin registration must be removed from that library since plugins are not allowed in regular libraries.